requirements.txt
# FastAPI and server
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database
supabase==2.0.2
psycopg[binary]

# Data processing
pandas==2.1.3
numpy==1.25.2
scikit-learn==1.3.2

# NLP and text processing
python-Levenshtein==0.27.1
rapidfuzz==3.14.0

# HTTP and networking
httpx==0.24.1
requests==2.31.0

# Environment and configuration
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Monitoring and logging
psutil==5.9.6

# Date and time
python-dateutil==2.8.2

# Security
cryptography==42.0.8

# Development and testing
pytest==7.4.3
pytest-asyncio==0.21.1

# Additional dependencies for production
aiofiles==23.2.0
typing-extensions==4.8.0
Pillow==10.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
email-validator==2.1.0
jinja2==3.1.2
python-magic==0.4.27
openpyxl==3.1.2
xlsxwriter==3.1.9
reportlab==4.0.7
qrcode[pil]==7.4.2
phonenumbers==8.13.25
geopy==2.4.0
timezonefinder==6.2.0
cachetools==5.3.2
redis==5.0.1
celery==5.3.4
flower==2.0.1
gunicorn==21.2.0
supervisor==4.2.5